pyproject.toml
[project]
name = "ceritificate"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "img2pdf>=0.5.1",
    "inquirer>=3.4.0",
    "pandas>=2.2.3",
    "pillow>=10.4.0",
    "requests>=2.32.3",
]
